게시판 페이지 업로드 구현중

+++ petmily/src/main/java/com/kh/petmily/entity/FaqFileDto.java
@@ -14,5 +14,5 @@ public class FaqFileDto {
 	private String uploadname;
 	private String savename;
 	private String filetype;
-	private int filesize;
+	private long filesize;
 }
